# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Pet Helper API Documentation
  version: 1.0.0
  description: Tài liệu API cho dự án Pet Helper. Hỗ trợ xác thực bằng JWT.
servers:
  - url: http://localhost:3000
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password, email]
              properties:
                username: { type: string }
                password: { type: string }
                email: { type: string }
                phone: { type: string }
      responses:
        201: { description: User registered successfully }
        400: { description: Invalid input }
  /api/v1/auth/login:
    post:
      summary: Login user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string }
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
        401: { description: Unauthorized }
...
  /api/v1/orders:
    post:
      summary: Create an order
      tags: [Orders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer: { type: object }
                items:
                  type: array
                  items: { type: object }
      responses:
        200: { description: Order created }
    get:
      summary: Get all orders (Admin)
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: List of all orders }
  /api/v1/orders/my:
    get:
      summary: Get my orders
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: List of my orders }
  /api/v1/orders/my/{id}/detail:
    get:
      summary: Get my order details
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Order details }
  /api/v1/orders/user/{user_id}:
    get:
      summary: Get orders by user id
      tags: [Orders]
      parameters:
        - in: path
          name: user_id
          required: true
          schema: { type: string }
      responses:
        200: { description: List of user orders }
  /api/v1/orders/{id}:
    put:
      summary: Update order status
      tags: [Orders]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string }
      responses:
        200: { description: Order updated }
  /api/v1/orders/admin/session/start:
    post:
      summary: Start admin session
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: Session started }

  # --- Products ---
  /api/v1/products:
    get:
      summary: Get all products
      tags: [Products]
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
      responses:
        200: { description: List of products }
    post:
      summary: Add product
      tags: [Products]
      security: [{ bearerAuth: [] }]
      responses:
        200: { description: Product added }
...
  /api/v1/reviews:
    post:
      summary: Add a review
      tags: [Reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id: { type: integer }
                user_id: { type: integer }
                rating: { type: integer }
                comment: { type: string }
      responses:
        200: { description: Review added }
...